fish: Do not change directory if Ctrl-C pressed in fuzzy search
In fish, running `cd foobar` will cd to the home directory (effectively running `cd` with no arguments) if the target directory does not exist.
The same behavior happens if `Ctrl-C` or `Esc` is pressed while the fuzzy search is open.

Related: #161 

This fix will make enhancd default to the working directory if no arguments are passed in, since that only happens when there is an error.

Fix suggested by @Gorthog in #161